fix: remove swagger tag
